Filtrex v0.0.1 Filtrex.AST

Filtrex.AST is a helper for building out the ecto macro query expression from a set of conditions. Typically, it shouldn’t need to be called direcly.

Functions

build_query(conditions, model, join)

Builds a ‘from’ ecto query from conditions and a join operator (e.g. ‘AND’)