Filtrex v0.2.0-dev Filtrex.Condition.Number

Filtrex.Condition.Number is a specific condition type for handling integer and decimal filters with various configuration options.

Comparators: is, is not, greater than, less than or, greater than or, less than

Configuation Options:

Key	Type	Description

allow_decimal	true/false	required to allow decimal values

allowed_values	list/range	value must be in these values

Types

t :: Filtrex.Condition.Number

Functions

comparators()

Callback implementation for Filtrex.Condition.comparators/0.

parse(config, map)

Callback implementation for Filtrex.Condition.parse/2.

type()

Callback implementation for Filtrex.Condition.type/0.