Filtrex v0.2.0 Filtrex.Type.Config

This configuration struct is for passing options at the top-level (e.g. Filtrex.parse/2) in a list.

Struct keys:

type: the corresponding condition module with this type (e.g. :text = Filtrex.Condition.Text)

keys: the allowed keys for this configuration

options: the configuration options to be passed to the condition

Types

t :: Filtrex.Type.Config

Functions

allowed?(configs, key)

Returns whether the passed key is listed in any of the configurations

config(configs, key)

Returns the configuration for the specified key

configs_for_type(configs, type)

Narrows the list of configurations to only the specified type

options(configs, key)

Returns the specific options of a configuration based on the key